+++ recipes/server.rb
@@ -16,12 +16,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+# Ensure common components are installed
 include_recipe "mcollective::common"
 
-package "mcollective" do
-  action :install
-  version node['mcollective']['package']['version']
-end
+# Install server components
+include_recipe node['mcollective']['recipes']['install_server']
 
 service "mcollective" do
   supports :restart => true, :status => true